Rename and moved converter to a more appropriate place.
